@@ -59,20 +59,20 @@
         <denodo-hibernate-dialect.version>8.0.20201021</denodo-hibernate-dialect.version>
         <denodo-jdbc.version>8.0.2</denodo-jdbc.version>
         <frontend-maven-plugin.version>1.11.0</frontend-maven-plugin.version>
-        <git-commit-id-maven-plugin.version>5.0.0</git-commit-id-maven-plugin.version>
-        <hibernate5-ddl-maven-plugin.version>2.5.0</hibernate5-ddl-maven-plugin.version>
-        <java.version>17</java.version>
-        <jdk.source>17</jdk.source>
-        <jdk.target>17</jdk.target>
-        <lms-embedded-services.version>5.2.14</lms-embedded-services.version>
-        <lms-team-spring-boot-it12>4.8</lms-team-spring-boot-it12>
-        <opencsv.version>5.8</opencsv.version>
-        <spring-cloud-starter-parent.version>2021.0.8</spring-cloud-starter-parent.version>
-        <springdoc-openapi-ui.version>1.7.0</springdoc-openapi-ui.version>
+        <git-commit-id-maven-plugin.version>9.0.1</git-commit-id-maven-plugin.version>
+        <hibernate6-ddl-maven-plugin.version>1.0.2</hibernate6-ddl-maven-plugin.version>
+        <java.version>21</java.version>
+        <jdk.source>21</jdk.source>
+        <jdk.target>21</jdk.target>
+        <lms-embedded-services.version>6.0.2-SNAPSHOT</lms-embedded-services.version>
+        <lms-team-spring-boot-it12>6.5.0</lms-team-spring-boot-it12>
+        <opencsv.version>5.9</opencsv.version>
+        <spring-cloud-starter-parent.version>2023.0.3</spring-cloud-starter-parent.version>
+        <springdoc-openapi-ui.version>2.6.0</springdoc-openapi-ui.version>
 
-        <plugins.compiler.version>3.11.0</plugins.compiler.version>
-        <plugins.license.version>2.0.0</plugins.license.version>
-        <plugins.release.version>3.0.0</plugins.release.version>
+        <plugins.compiler.version>3.13.0</plugins.compiler.version>
+        <plugins.license.version>2.4.0</plugins.license.version>
+        <plugins.release.version>3.1.1</plugins.release.version>
     </properties>
 
     <dependencyManagement>
