@@ -35,13 +35,13 @@ package edu.iu.uits.lms.sisgradesexport.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import edu.iu.uits.lms.canvas.model.User;
+import jakarta.xml.bind.annotation.XmlAccessType;
+import jakarta.xml.bind.annotation.XmlAccessorType;
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlRootElement;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 import java.io.Serializable;
 
 @JsonIgnoreProperties (ignoreUnknown=true)
