@@ -33,9 +33,9 @@ package edu.iu.uits.lms.sisgradesexport.services;
  * #L%
  */
 
-import edu.iu.uits.lms.iuonly.services.SisServiceImpl;
 import edu.iu.uits.lms.lti.config.TestUtils;
-import edu.iu.uits.lms.sisgradesexport.config.ToolConfig;
+import edu.iu.uits.lms.lti.service.LmsDefaultGrantedAuthoritiesMapper;
+import edu.iu.uits.lms.sisgradesexport.config.SecurityConfig;
 import edu.iu.uits.lms.sisgradesexport.controller.rest.SisGradesRestController;
 import edu.iu.uits.lms.sisgradesexport.service.SisGradesExportService;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -44,15 +44,16 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.Instant;
@@ -64,7 +65,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(value = SisGradesRestController.class, properties = {"oauth.tokenprovider.url=http://foo"})
-@Import(ToolConfig.class)
+@ContextConfiguration(classes = {SisGradesRestController.class, SecurityConfig.class})
 @Slf4j
 @ActiveProfiles("none")
 public class RestLaunchSecurityTest {
@@ -75,7 +76,10 @@ public class RestLaunchSecurityTest {
    private SisGradesExportService sisGradesExportService;
 
    @MockBean
-   private SisServiceImpl sisService;
+   private LmsDefaultGrantedAuthoritiesMapper lmsDefaultGrantedAuthoritiesMapper;
+
+   @MockBean
+   private ClientRegistrationRepository clientRegistrationRepository;
 
    @Test
    public void restNoAuthnLaunch() throws Exception {
